% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{build_UDF_list}
\alias{build_UDF_list}
\title{Build a list of User Defined Fields (UDFs) for query}
\usage{
build_UDF_list(addl_columns = NULL, include_defaults = TRUE)
}
\arguments{
\item{addl_columns}{A character vector of valid UDF SQL headers.}

\item{include_defaults}{A boolean indicator of whether to include the default
UDFs of "FreezerName", "Position1", "Position2", and "Position3".}
}
\value{
A character scalar of UDFs formatted properly.
}
\description{
This function formats UDFs for use with SQL queries. There is no support or
checks for invalid headers. This is intended as an internal support function
only. It will always include Globally Unique Aliquot ID (GUALIQUOTID).
}
\examples{
build_UDF_list()
build_UDF_list(c("Field_ID_Num", "Project_ID"))
build_UDF_list(c("Field_ID_Num", FALSE))
}
