% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{freezercheck_demo}
\alias{freezercheck_demo}
\title{Demonstration of Freezercheck}
\usage{
freezercheck_demo()
}
\description{
Literally only a wrapper to start Freezercheck in demonstration mode.
}
\examples{
freezercheck_demo()
}
