% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{freezercheck_setup}
\alias{freezercheck_setup}
\title{Set up FreezerCheck}
\usage{
freezercheck_setup(overwrite_demo = TRUE, overwrite_data = FALSE)
}
\arguments{
\item{overwrite_demo}{Logical indicator to overwrite existing files in the
demonstration directory. Defaults to TRUE.}

\item{overwrite_data}{Logical indicator to overwrite existing files in the
user-defined data directory. Defaults to FALSE.}
}
\description{
Does the heavy lifting of setting up your system to work with your
installation of Freezerworks and the Freezerworks ODBC client. This should
only need to be run on package installation, but can also be used to reset
your version of FreezerCheck. If run again after installation, by default
will not overwrite existing data in the user directory, but will reset any
data modified during demo mode. This can also be used to reset your
Freezerworks DSN assignment in cases of updated or renamed DSNs. Multiple
installations of FreezerCheck are supported via alias. When launching from
the console, you will be prompted for which to use.
}
