% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{freezercheck_start}
\alias{freezercheck_start}
\title{Start the FreezerCheck quality assurance dashboard}
\usage{
freezercheck_start(demo_mode = FALSE)
}
\arguments{
\item{demo_mode}{A boolean indicator (default FALSE) of whether to launch
FreezerCheck in demonstration mode.}
}
\description{
Use this to launch FreezerCheck from the console. Directory must be in place.
}
\examples{
start_freezercheck(demo_mode = TRUE)
}
