% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_fields}
\alias{get_fields}
\title{Get the list of fields from Freezerworks}
\usage{
get_fields(DSN, dir_user_path = NULL)
}
\arguments{
\item{DSN}{Character scalar of the data source name for your Freezerworks
ODBC driver.}

\item{dir_user_path}{Character scalar of the data path to your user
directory, only use to write back any changes (defaults to NULL).}
}
\value{
A character vector of UDFs
}
\description{
This function creates the list of valid User Defined Fields (UDFs) used in
your Freezerworks tables 'Aliquots' and 'Samples'. Works with demo mode. If
demo mode is off, ensure the DSN is valid first. Parameter dir_user_path is
only necessary if you wish to overwrite the current list of fields. Edit the
object at dir_user_path/fields.RDS to customize this list for use with
FreezerCheck and only include a subset of UDFs for audit purposes. By default
this will return EVERY UDF present, which is likely not suitable.
}
\examples{
demo_mode <- TRUE
DSN <- "ODBC_Name"
get_fields(DSN)
}
