% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_freezers}
\alias{get_freezers}
\title{Get the list of freezers from Freezerworks}
\usage{
get_freezers(DSN = NULL, dir_user_path = NULL)
}
\arguments{
\item{DSN}{Character scalar of the data source name for your Freezerworks
ODBC driver.}

\item{dir_user_path}{Character scalar of the data path to your user
directory, only use to write back any changes.}
}
\value{
A character vector of freezers.
}
\description{
This function creates the list of valid freezers for use with FreezerCheck.
Works with demo mode. Ensure DSN is set in your environment.
}
\examples{
demo_mode <- TRUE
get_freezers(DSN)
}
