% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_GUAID_info}
\alias{get_GUAID_info}
\title{Get additional information about an aliquot}
\usage{
get_GUAID_info(GUAID, addl_columns = NULL, include_defaults = TRUE)
}
\arguments{
\item{GUAID}{A character vector of GUAIDs as stored in Freezerworks.}

\item{addl_columns}{A character vector of columns to include other than
defaults that will be passed to build_UDF_list().}

\item{include_defaults}{A boolean indicator of whether to include the default
UDFs of "FreezerName", "Position1", "Position2", and "Position3" that will
be passed to build_UDF_list().}

\item{demo_mode}{A boolean indicator to use demo data rather than a live DSN
connection.}
}
\value{
Assuming the output of build_UDF_list() is valid, a data frame is
  returned containing the requested columns from Freezerworks with one row
  for each GUAID provided.
}
\description{
This queries Freezerworks with a list of globally unique aliquot IDs (GUAIDs)
and returns additional metadata to assist with the audit process. In order to
work, this must include a valid data source name to connect with
Freezerworks.
}
\examples{
get_GUAID_info(GUAID = "DEM204978O", demo_mode = TRUE)
get_GUAID_info(GUAID = "DEM218508O", addl_columns = "PROJECT_ID", demo_mode = TRUE)
get_GUAID_info(GUAID = c("DEM137314O", "DEM220303O", "DEM122985O"), demo_mode = TRUE)
}
