% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_scan_positions}
\alias{get_scan_positions}
\title{Get positions to audit}
\usage{
get_scan_positions(dsn, freezer, addl_columns = NULL, fraction = 0.1,
  return_all = FALSE, random = FALSE, max_returns = NA,
  focus_racks = TRUE, condense_rack_to = 0.33,
  condense_box_to = 0.25)
}
\arguments{
\item{dsn}{Character scalar for the data source name to connect to
FreezerWorks.}

\item{freezer}{Character scalar with the freezer being targeted for audit.}

\item{addl_columns}{Character vector of any additional User Defined Fields to
include. Failure will occur if these do not exist.}

\item{fraction}{A umerical scalar, default 0.1, of the fraction of positions
in the freezer to audit from 0 to 1.}

\item{return_all}{Logical indicator, default FALSE, to trigger a full audit.}

\item{random}{Logical indicator, default FALSE, to Obtain a true random
selection of positions.}

\item{max_returns}{Numerical scalar indicating the absolute maximum number of
positions to return.}

\item{focus_racks}{Logical indicator, default TRUE, of whether to perform
focused condensing for less effort.}

\item{condense_rack_to}{Numerical scalar from 0 to 1, default 0.33, of the
target fraction of first order containers - e.g. freezer racks - to
condense. For example, if there are 10 boxes in a rack, 0.33 will return 3
of them for audit.}

\item{condense_box_to}{Numerical scalar from 0 to 1, default 0.25, of the
target fraction of second order containers - e.g. boxes within racks - to
condense. For example, if there are 81 places in a box, 0.25 will return 21
of them for audit.}

\item{demo_mode}{Logical indicator, default FALSE, on whether to run in demo
mode. Allows for no external database connection and demonstrates
functionality only.}
}
\value{
A data frame of aliquot positions to audit.
}
\description{
Obtains a random sampling of positions to audit from a specific freezer
within the repository managed by FreezerWorks. Effort optimization is done by
condensing structured containers by iterative amounts.
}
\examples{
demo_mode <- TRUE
get_scan_positions(freezer = "Freezer 1")
}
