% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_discrep}
\alias{make_discrep}
\title{Convenience function to properly format the discrepancies object}
\usage{
make_discrep()
}
\value{
A mutated version of the discrepancies data frame.
}
\description{
This is purely an internal function to join the discrepancies and audits
tables during launch of FreezerCheck.
}
