% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_progress}
\alias{make_progress}
\title{Create the progress data frame.}
\usage{
make_progress(freezers)
}
\arguments{
\item{freezers}{Character vector of freezers being tracked.}
}
\value{
A data frame formatted for graphing.
}
\description{
Returns the object necessary to properly plot freezer progress by assigning x
and y coordinates and baseline progress value.
}
\examples{
demo_mode <- TRUE
freezer_list <- get_freezers()
progress <- make_progress(freezer_list)
}
