% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{next_audit_index}
\alias{next_audit_index}
\title{Get the next audit index}
\usage{
next_audit_index(index, this_audit)
}
\arguments{
\item{index}{Integer scalar of the current audit aliquot index.}

\item{this_audit}{Data frame of the current audit.}
}
\value{
Integer scalar of the next audit aliquot index.
}
\description{
This selectively chooses the next audit index when a position is either
verified or rejected. By default it selects the next pending position check
in line, but also catches overflow and wraps back to prior positions or
returns 0 if all positions have been checked. This is intended as an internal
function only.
}
