% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parse_container}
\alias{parse_container}
\title{Create meaningful descriptors of containers}
\usage{
parse_container(section)
}
\arguments{
\item{section}{The name of a freezer section from Freezerworks.}
}
\value{
A list of the prefixes for Position1, Position2, and Position3.
}
\description{
This leverages the Freezerworks Section name (header 'FreezerName') to decide
on meaningful prefixes corresponding to values in Position1, Position2, and
Position3. This is intended as an internal function only, and heavily slanted
toward the use of Freezerworks within the NIST Biorepository. Feel free to
modify for your use to customize display options given your Freezerworks
setup. This is run on a section-by-section basis and the return is used to
modify those columns for tabular display within FreezerCheck.
}
