% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simple_cap}
\alias{simple_cap}
\title{Clean up table headers}
\usage{
simple_cap(s, sep = ".")
}
\arguments{
\item{s}{Character vector of words to clean.}

\item{sep}{Character scalar indicating word separation character. (Default
"." but most used as "_".)}
}
\value{
Character vector matching parameter s with clean formatting.
}
\description{
Pulling data from Freezerworks often mangles header case presentation and
returns SQL names instead of display names. This provides sentence-case
formatting using a user-defined separator. As written, provides some
additional cleanup of commonly mis-formatted headers used in the NIST
Biorepository. Feel free to modify the list for your needs.
}
\examples{
if (!exists("demo_data")) demo_data <- read_rds("data/demo/demo_data.RDS")
simple_cap(names(demo_data, sep = "_"))
}
