% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{view_audit}
\alias{view_audit}
\title{View a given audit as a DT datatable}
\usage{
view_audit(audit)
}
\arguments{
\item{audit}{Data frame of the audit to display}
}
\value{
DT::datatable object of specific type
}
\description{
This internal helper function does nothing more than consistently create a
DT::datatable() object from a given audit.
}
\examples{
# Must have audits present to be valid
view_audit(subset(audits, audit_index == 1))
}
