% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{view_discrep_details}
\alias{view_discrep_details}
\title{Describe a discrepancy}
\usage{
view_discrep_details(discrep_index)
}
\arguments{
\item{discrep_index}{Integer scalar index value identifying a discrepancy.}
}
\value{
HTML snippet
}
\description{
Creates the HTML to populate a modal display for describing a particular
discrepancy. This is intended as an internal function only.
}
